More refactoring fun :)

+++ src/BeHeroes.DigitalTwins.Core/Synchronization/ISequencer.cs
@@ -1,22 +1,24 @@
+using System.Numerics;
+
 namespace BeHeroes.DigitalTwins.Core.Synchronization
 {
     //TODO: Migrate to BeHeroes.CodeOps.Abstractions package in Numerics namespace
     /// <summary>
-    /// Defines an interface for a sequencer that generates a sequence of unique seeds.
+    /// Defines an interface for a sequencer that generates a sequence of incrementing values.
     /// </summary>
     public interface ISequencer
     {
         /// <summary>
-        /// Gets the current from the sequence.
+        /// Gets the current value from the sequence.
         /// </summary>
         /// <returns>The current seed.</returns>
-        long? Current();
+        BigInteger? Current();
         
         /// <summary>
         /// Gets the next value from the sequence.
         /// </summary>
         /// <returns>The next seed.</returns>
-        long Next();
+        BigInteger Next();
 
         /// <summary>
         /// Resets the sequencer.
