+++ src/BeHeroes.DigitalTwins.Core/Synchronization/StateShadow.cs
@@ -1,16 +1,18 @@
+using System.Numerics;
+
 namespace BeHeroes.DigitalTwins.Core.Synchronization
 {
     /// <summary>
-    /// Represents the shadow state of a digital twin.
+    /// Represents the shadow state of a replica.
     /// </summary>
     public sealed class StateShadow : State, IStateShadow
     {
-        private ulong peerVersion = default!;
+        private BigInteger peerVersion = default!;
 
         /// <summary>
         /// Gets or sets the version number of the peer.
         /// </summary>
-        public ulong PeerVersion { get => peerVersion; set => peerVersion = value; }
+        public BigInteger PeerVersion { get => peerVersion; set => peerVersion = value; }
         
         /// <summary>
         /// Initializes a new instance of the <see cref="StateShadow"/> class.
@@ -18,7 +20,7 @@ namespace BeHeroes.DigitalTwins.Core.Synchronization
         /// <param name="data">The current state data.</param>
         /// <param name="version">The version of the current state data.</param>
         /// <param name="previousData">The previous state data.</param>
-        public StateShadow(object data, ulong version, object? previousData) : base(data, version, previousData)
+        public StateShadow(object data, BigInteger version, object? previousData) : base(data, version, previousData)
         {
         }
     }
